refactor: export Education types with aliases from types.ts file
